requirements.txt
python-telegram-bot==9.0.0
requests==2.18.4